[Store] Extract shared MasterServiceTest fixture; split group/lease tests into master_service_group_test
